+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- added the `config.tui.scroll_offset` setting
+
+### Fixed
+- retain scroll position in the TUI's list view
+
 
 ## [4.1.0] - 2023-06-11